% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataimportModule.R
\name{dataimportUI}
\alias{dataimportUI}
\title{Data import module UI}
\usage{
dataimportUI(id = "dataimport", fileExt = SUPPORTED_FILES,
  label = "Choose a file")
}
\description{
Data import module UI
}

